more fixings

+++ ajm/generate.py
@@ -1,6 +1,6 @@
 from textwrap import dedent
-from .helpers import warning_text, success_text
-
+from .helpers import warning_text, success_text, debug_text
+from pathlib import Path
 from jinja2 import Environment, FileSystemLoader
 
 CLOUDBUILD_CONFIG = "_cloudbuild.yaml"
@@ -10,13 +10,13 @@ environment = Environment(loader=FileSystemLoader("ajm/templates/"))
 template = environment.get_template("cloudbuild.yaml.tmpl")
 
 
-def generate_cloudbuildyaml(settings):
-    settings["region"] = "us-central1"  # TODO: make dynamic.
-
+def generate_cloudbuildyaml(settings, region):
+    settings["region"] = region
     content = template.render(**settings)
 
     # pyyaml loses comments https://github.com/yaml/pyyaml/issues/90
     # more correct validation would be this method, but we lose fidelity.
+    # 
     # validate = yaml.safe_load(content)
     # return yaml.dump(validate, sort_keys=False)
 
@@ -34,7 +34,7 @@ def generate_tfvars(repo, region):
         cloudbuild_file = "{CLOUDBUILD_CONFIG}"
 
         # Generate at https://github.com/settings/tokens/new
-        # Use Classic Topken with repo and read:user permissions"
+        # Use Classic Token with repo and read:user permissions"
         github_token = ""
         
         # ID from "Cloud Build" app on https://github.com/settings/installations
@@ -42,9 +42,14 @@ def generate_tfvars(repo, region):
     """
     )
 
-    warning_text(
-        f"Update the github_token and installation_id values in {TFVARS_CONFIG}"
-    )
+    if not Path(TFVARS_CONFIG).exists(): 
 
-    open(TFVARS_CONFIG, "w").write(tfvars)
-    success_text(f"Wrote Terraform variables to {TFVARS_CONFIG}")
+        open(TFVARS_CONFIG, "w").write(tfvars)
+        success_text(f"Wrote Terraform variables to {TFVARS_CONFIG}")
+
+        warning_text(
+            f"Update the github_token and installation_id values in {TFVARS_CONFIG}"
+        )
+
+    else: 
+        debug_text(f"Using existing {TFVARS_CONFIG}.")
